+++ src/setup/Input.ts
@@ -1,0 +1,36 @@
+import { applyDefaults, cn } from '../../deps.ts';
+import { iInput } from '../types/props.ts';
+import { boxInput, nonboxInput, button } from './shared.ts';
+
+const inputDefaults: iInput = {
+  required: false,
+};
+
+export default (props: iInput) => {
+  const p = applyDefaults<iInput>(inputDefaults, props);
+
+  const inputClassnames = {
+    input: cn(
+      // Applies to all inputs
+      'outline-cobalto',
+      // Box types vs non-box types vs button types
+      p.type && ['button', 'image', 'reset', 'submit'].includes(p.type)
+        ? button(p.disabled, p.error)
+        : p.type === 'radio' || p.type === 'checkbox' || p.type === 'slider'
+        ? nonboxInput(p.disabled)
+        : boxInput(p.disabled, p.error)
+    ),
+    span: cn('px-2 text-obsidiana w-full'),
+    label: cn(
+      'flex',
+      p.type === 'radio' || p.type === 'checkbox'
+        ? 'flex-row-reverse justify-end'
+        : 'flex-col'
+    ),
+    error: cn('px-2 text-xs text-chicle'),
+    required: cn('text-chicle ml-1 mb-1'),
+    container: cn('input'),
+  };
+
+  return { c: inputClassnames, ...p };
+};